SIREHC/requirements.txt
customtkinter==5.2.2
pillow==12.0.0
openai==2.7.2
pandas==2.3.3
matplotlib==3.10.7
requests==2.32.5
numpy==2.3.4
